[equipped-item-search] move : api response 패키지 분리
