+++ core/data/src/main/java/com/hegunhee/maplefinder/data/api/MapleCharacterApi.kt
@@ -1,10 +1,10 @@
 package com.hegunhee.maplefinder.data.api
 
-import com.hegunhee.maplefinder.data.api.model.character.CharacterAbilityResponse
+import com.hegunhee.maplefinder.data.api.model.character.stat.CharacterAbilityResponse
 import com.hegunhee.maplefinder.data.api.model.character.CharacterDojangResponse
 import com.hegunhee.maplefinder.data.api.model.character.CharacterBasicResponse
-import com.hegunhee.maplefinder.data.api.model.character.CharacterHyperStatResponse
-import com.hegunhee.maplefinder.data.api.model.character.CharacterStatResponse
+import com.hegunhee.maplefinder.data.api.model.character.stat.CharacterHyperStatResponse
+import com.hegunhee.maplefinder.data.api.model.character.stat.CharacterStatResponse
 import retrofit2.http.GET
 import retrofit2.http.Query
